fix(playground): move constant
